+++ packages/orq-rc/src/orq_ai_sdk/models/filelistop.py
@@ -74,7 +74,7 @@ class FileListData(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-09-05T09:48:59.769Z")
+    created: Optional[datetime] = parse_datetime("2025-09-05T11:12:44.839Z")
     r"""The date and time the resource was created"""