+++ packages/orq-rc/RELEASES.md
@@ -2659,3 +2659,13 @@ Based on:
 - [python v3.12.0-rc.42] packages/orq-rc
 ### Releases
 - [PyPI v3.12.0-rc.42] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.42 - packages/orq-rc
+
+## 2025-09-05 11:14:19
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.609.0 (2.692.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.12.0-rc.43] packages/orq-rc
+### Releases
+- [PyPI v3.12.0-rc.43] https://pypi.org/project/orq-ai-sdk/3.12.0-rc.43 - packages/orq-rc